IQuantity with QuantityValue as Value type (PoC) - QuantityValue implements most of the INumber interface members / operators (operations  preserving decimal when possible, failing back to double, clamped to the real values- infinities excluded) - changed the type of IQuantity.Value/As to QuantityValue (explicit cast required) - added back the IEquatable interface: behavior depending on the runtime type of the quantity values (strict equality comparison for double/double) - added back the fix for the non-reflexive CompareTo issue (and the corresponding tests) - added a check preventing the possible stack overflow exception when calling ToUnit() with default(TUnit) (or any other invalid Enum value) - DataContracts changed to reflect the possible use of either double or decimal as ValueType (the Json converters were not changed- but probably should be) - removed the obsolete Json converter tests

+++ UnitsNet.Tests/CustomCode/PowerRatioTests.cs
@@ -41,11 +41,12 @@ namespace UnitsNet.Tests.CustomCode
         public void ExpectPowerConvertedCorrectly(double power, double expected)
         {
             Power p = Power.FromWatts(power);
-            double actual = PowerRatio.FromPower(p).DecibelWatts;
+            double actual = (double)PowerRatio.FromPower(p).DecibelWatts;
             Assert.Equal(expected, actual);
         }
 
         [Theory]
+        // Note: Attribute arguments cannot be of type decimal.
         [InlineData(-20, 0.01)]
         [InlineData(-10, 0.1)]
         [InlineData(0, 1)]
@@ -54,8 +55,8 @@ namespace UnitsNet.Tests.CustomCode
         public void ExpectPowerRatioConvertedCorrectly(double powerRatio, double expected)
         {
             PowerRatio pr = PowerRatio.FromDecibelWatts(powerRatio);
-            double actual = pr.ToPower().Watts;
-            Assert.Equal(expected, actual);
+            decimal actual = (decimal)pr.ToPower().Watts;
+            Assert.Equal((decimal)expected, actual);
         }
 
         // http://www.maximintegrated.com/en/app-notes/index.mvp/id/808
@@ -69,7 +70,7 @@ namespace UnitsNet.Tests.CustomCode
         {
             PowerRatio powerRatio = PowerRatio.FromDecibelMilliwatts(dBmW);
 
-            double actual = Math.Round(powerRatio.ToAmplitudeRatio(ElectricResistance.FromOhms(50)).DecibelMillivolts, 2);
+            double actual = Math.Round((double)powerRatio.ToAmplitudeRatio(ElectricResistance.FromOhms(50)).DecibelMillivolts, 2);
             Assert.Equal(expected, actual);
         }
 
@@ -82,7 +83,7 @@ namespace UnitsNet.Tests.CustomCode
         {
             PowerRatio powerRatio = PowerRatio.FromDecibelMilliwatts(dBmW);
 
-            double actual = Math.Round(powerRatio.ToAmplitudeRatio(ElectricResistance.FromOhms(75)).DecibelMillivolts, 2);
+            double actual = Math.Round((double)powerRatio.ToAmplitudeRatio(ElectricResistance.FromOhms(75)).DecibelMillivolts, 2);
             Assert.Equal(expected, actual);
         }
     }
